apps/backend/requirements.txt
djangorestframework
django-cors-headers
prisma-client-py
